Move handlers to beta
